Juntar componentes de lista de consumo
