--- app/src/main/java/br/com/ecowatt/ui/components/consumption/ListItem.kt
@@ -1,123 +1,0 @@
-package br.com.ecowatt.ui.components.consumption
-
-import androidx.compose.foundation.clickable
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
-import androidx.compose.material3.ListItem
-import androidx.compose.material3.ListItemColors
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.tooling.preview.Preview
-import br.com.ecowatt.R
-import br.com.ecowatt.models.Device
-import br.com.ecowatt.sampledata.highConsumptionLevelDeviceSampleData
-import br.com.ecowatt.ui.theme.Gray400
-import br.com.ecowatt.ui.theme.Gray600
-import br.com.ecowatt.ui.theme.Gray800
-import br.com.ecowatt.ui.theme.Neutral1000
-import br.com.ecowatt.ui.theme.Red100
-import br.com.ecowatt.ui.theme.Red400
-
-/**
- * Component to display an energy consumption list item.
- *
- * @param device the device whose energy consumption is being displayed.
- * @param isConsumptionHigh indicates if the device's energy consumption is high.
- * @param onClick called when this list item is clicked.
- * @param onDeleteDevice called when the delete button of this list item is clicked.
- * @see Device
- * @see EnergyConsumptionList
- */
-@Composable
-internal fun EnergyConsumptionListItem(
-    device: Device,
-    isConsumptionHigh: Boolean,
-    onClick: () -> Unit,
-    onDeleteDevice: () -> Unit
-) {
-    ListItem(
-        modifier = Modifier.clickable { onClick() },
-        colors = getListItemColors(isConsumptionHigh),
-        leadingContent = {
-            ConsumptionLevelIcon(isConsumptionHigh = isConsumptionHigh)
-        },
-        headlineContent = {
-            Text(device.name)
-        },
-        supportingContent = {
-            Text("${device.getCurrentConsumptionLevel()} watts/minute")
-        },
-        trailingContent = {
-            IconButton(onClick = onDeleteDevice) {
-                Icon(
-                    painter = painterResource(R.drawable.ic_delete),
-                    contentDescription = "Delete device"
-                )
-            }
-        }
-    )
-}
-
-/**
- * Function to get the colors for the list item based on consumption level.
- *
- * @param isConsumptionHigh indicates if the device's energy consumption is high.
- * @return [ListItemColors] the colors to be used for the list item.
- * @see Device
- * @see EnergyConsumptionListItem
- * @see ListItemColors
- */
-private fun getListItemColors(
-    isConsumptionHigh: Boolean
-): ListItemColors {
-    val listItemColorsLowConsumption = ListItemColors(
-        containerColor = Neutral1000,
-        headlineColor = Gray800,
-        leadingIconColor = Gray600,
-        overlineColor = Gray600,
-        supportingTextColor = Gray600,
-        trailingIconColor = Gray600,
-        disabledHeadlineColor = Gray600,
-        disabledLeadingIconColor = Gray400,
-        disabledTrailingIconColor = Gray400
-    )
-
-    val listItemColorsHighConsumption = ListItemColors(
-        containerColor = Red100,
-        headlineColor = Red400,
-        leadingIconColor = Red400,
-        overlineColor = Red400,
-        supportingTextColor = Red400,
-        trailingIconColor = Red400,
-        disabledHeadlineColor = Red400,
-        disabledLeadingIconColor = Red400,
-        disabledTrailingIconColor = Red400
-    )
-
-    return if (isConsumptionHigh) listItemColorsHighConsumption
-    else listItemColorsLowConsumption
-}
-
-@Preview
-@Composable
-private fun ListItemLowLevelPreview() {
-    EnergyConsumptionListItem(
-        device = highConsumptionLevelDeviceSampleData,
-        isConsumptionHigh = false,
-        onClick = {},
-        onDeleteDevice = {}
-    )
-}
-
-@Preview
-@Composable
-private fun ListItemHighLevelPreview() {
-    EnergyConsumptionListItem(
-        device = highConsumptionLevelDeviceSampleData,
-        isConsumptionHigh = true,
-        onClick = {},
-        onDeleteDevice = {}
-    )
-}
